Refactor out `lovatAPI.setSourceTeams()` and `lovatAPI.getSourceTeamSettings()`
