+++ lib/pages/onboarding/onboarding_page.dart
@@ -14,6 +14,7 @@ import 'package:scouting_dashboard_app/reusable/lovat_api/onboarding/join_team_b
 import 'package:scouting_dashboard_app/reusable/lovat_api/onboarding/register_team.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/onboarding/registration_status.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/set_username.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/source_data/source_teams.dart';
 import 'package:scouting_dashboard_app/reusable/models/team.dart';
 import 'package:scouting_dashboard_app/reusable/page_body.dart';
 import 'package:shared_preferences/shared_preferences.dart';
